Didn't make much progress
I changed maybe a few things? Nothing much at all, I was trying to get
the next piece you choose to appear after placing the first one with the
current system...we may need an extra bool somewhere to keep track of
things, but the multiple listeners makes it kind of tough to follow.
Sorry guys :/