Game is up and running, server is communicating
and we can play multiplayer. We still need to add rotation and flip
features which are going to be hard. Because when we send the state of
the image we don't send the image rotated so we need to come up with a
solution. Also need to rethink checking end of game scenario because we
also need to check if player can make move with a rotated piece not just
current piece shape. Also thinking of making a local version of the game
option that doesn't require server but we will do this later.