something about the database Ill add it **
RYAN can you please write a function in the database that will just
output all of the data, like a select * function and just have it print
all the stats. we wont run it in anything we will just have another main
in the database but just so we can see what the stats are without having
to change them. cause right now the server will always update if it
loads someones information. cause my logic is once you join you are
there for the entire game regardless if you quit