trying to get jars to work completely...
hey so i just made the client login without checking database cause i
thought that caused our jars to exit but its not only that it is
something else in the Blokus file. so kyle try and look around a littlle
bit in there tmrw. I need to go to bed.